- Added sequential Kalman filter with Cholesky decomposition - Inversion seems to work

+++ data_processing.h
@@ -1,7 +1,7 @@
 #ifndef __MATH_UTILS_INCLUDED__
 #define __MATH_UTILS_INCLUDED__
 
-int data_inversion(char *input_file, char *output_file, double *args);
+int data_inversion(char *input_file, char *output_file, long double *args);
 
 #endif
